Create dualmonitorXPS13.sh
A basic script to get working two monitors and touchscreen on hidpi xps13 2016 with Xorg. It scales the second monitor and restart plasma.  The second monitor appears with a lof of blurry text but with scale to 1.6 and DPI to 158 I got a good compromise to work with two screens.